Get test clients to choose a server via a function.

+++ src/test_recs.h
@@ -14,7 +14,7 @@
 
 
 using ClientId = std::string;
-using ServerId = double;
+using ServerId = int;
 
 
 struct TestRequest {
